pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panda_lib"
version = "2.0"
description = "A library for operating a PANDA SDL as a standalone device."
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = { text = "GPL-2.0" }
authors = [
    { name = "Gregory Robben", email = "grobben@bu.edu" },
    { name = "Harley Quinn", email = "hjquinn@bu.edu" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
]
dependencies = [
    "annotated-types==0.7.0",
    "comtypes==1.4.8",
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "filelock==3.16.0",
    "fonttools==4.60.2",
    "fsspec==2024.9.0",
    "gitdb==4.0.11",
    "GitPython==3.1.43",
    "gpytorch==1.13",
    "greenlet==3.1.0",
    "hardpotato==1.4.0",
    "imageio==2.35.1",
    "jaxtyping==0.2.19",
    "Jinja2==3.1.6",
    "joblib==1.4.2",
    "kiwisolver==1.4.7",
    "lazy_loader==0.4",
    "linear-operator==0.5.3",
    "MarkupSafe==2.1.5",
    "matplotlib==3.9.2",
    "mpmath==1.3.0",
    "nesp-lib==0.2.0",
    "networkx==3.3",
    "numpy==2.2.5",
    "opencv-python==4.10.0.84",
    "packaging==24.1",
    "pandas==2.2.2",
    "pillow==10.4.0",
    # "pip>=25.3",  # REMOVED: Package manager, not a dependency
    "pre-commit>=4.2.0",
    "PuLP==2.9.0",
    "pydantic==2.9.1",
    "pydantic_core==2.23.3",
    "pyparsing==3.1.4",
    "pyserial==3.5",
    "pyserial-asyncio==0.6.0",
    "pytest>=8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "regex==2024.7.24",
    "ruff>=0.8.3",
    "sartorius==0.5.1",
    "scikit-image==0.24.0",
    "scikit-learn==1.5.1",
    "scipy>=1.14.1",
    "six==1.16.0",
    "slack_sdk>=3.4.0",
    "smmap==5.0.1",
    "SQLAlchemy==2.0.34",
    "sympy==1.13.2",
    "threadpoolctl==3.5.0",
    "tifffile==2024.8.30",
    "tomli==2.0.1",
    "torch>=2.7.1-rc1",
    "torchaudio>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm==4.67.1",
    "typeguard==4.3.0",
    "typing_extensions==4.12.2",
    "tzdata==2024.1",
    # NOTE: pip and uv are package managers, not library dependencies
    # They should be installed separately, not as dependencies
    # "pip>=25.3",  # REMOVED: Package manager, not a dependency
    # "uv>=0.8.6",  # REMOVED: Package manager, not a dependency
    "simple-pyspin==0.1.1",
    "websocket-client==1.8.0",
    # UNCOMMENT THE LINE BELOW WHEN YOU HAVE THE SPINNKER WHL FILE 
    # "spinnaker-python",
    "alembic",
    "mako"
]

[project.urls]
"Homepage" = "https://github.com/BU-KABlab/PANDA-BEAR"
"Bug Tracker" = "https://github.com/BU-KABlab/PANDA-BEAR/issues"
[project.optional-dependencies]
flir = ["simple-pyspin==0.1.1"]                           
syringe_pump = ["nesp_lib>=0.2.0"]
dev = ["pytest>=8.3.4", "ruff>=0.8.3", "mypy>=1.8.0", "black>=24.3.0"]

[project.scripts]
panda-cli = "panda_lib_cli.main:main"
panda-db-setup = "panda_lib_db.cli:main"
panda-slack-bot = "panda_lib.slack_tools.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"panda_lib.hardware.mill" = ["*.json", "*.txt"]
"panda_shared.config" = ["*.ini"]

[tool.pytest.ini_options]
pythonpath = "src"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "nesp_lib",
    "sartorius",
]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.uv.sources]
panda_sdl = { workspace = true }
panda-lib = { workspace = true }
hardpotato = { git = "https://github.com/BU-KABlab/hardpotato.git" }
nesp-lib = { git = "https://github.com/BU-KABlab/nesp-lib-py.git" }
# UNCOMMENT THE LINE BELOW WHEN YOU HAVE THE SPINNKER WHL FILE 
# spinnaker-python = { path = "../../spinnaker_python-4.2.0.46-cp310-cp310-linux_aarch64.whl" }

[dependency-groups]
dev = ["panda-lib"]
